Schedule saving complete, add pcb images
Most of the UI elements behave as expected, but setting the Running
state is not yet complete. Still need to work out how to deal with the
profile names on the server side. Added a couple of images for the PCB.
Rendered version of the pcb is care of dirtypcb.com, where I ordered the
first batch of PCBs. The other image is rendered via EagleCAD.